// src/app/backend/types/models/entity.d.ts
type GetClient = Prisma.ClientGetPayload<{
  select: {
    id: true;
    fullName: true;
    fullSurname: true;
    identified: true;
    clientState: true;
    clientContact: {
      select: {
        phoneNumber: true;
        email: true;
        address: true;
      };
    };
    clientVehicle: {
      select: {
        brand: true;
        model: true;
        year: true;
        plates: true;
        engineDisplacement: true;
        description: true;
      };
    };
  };
}>;

type UpdateClient = Record<{
  clientState: boolean;
  phoneNumber: string;
  email: string;
  address: string;
  description?: string;
}>;

type CreateClient = {
  fullName: string;
  fullSurname: string;
  identified: string;
  clientContact: {
    phoneNumber: string;
    email: string;
    address?: string;
  };
  clientVehicle?: [
    {
      brand: string;
      model: string;
      year: number;
      plates: string;
      engineDisplacement: number;
      description?: string;
    }
  ];
};